src/api/requirements.txt
Flask>=2.0
Flask-Admin>=1.5
Flask-Migrate>=3.0
Flask-SQLAlchemy>=3.0
Flask-Swagger>=0.2
Flask-JWT-Extended>=4.0
Flask-Bcrypt>=1.0
Flask-Cors>=3.0
python-dotenv>=1.0
stripe>=5.0
click>=8.0
SQLAlchemy>=1.4
itsdangerous>=2.0
gunicorn>=20.0
psycopg2-binary>=2.9
twilio>=8.0
openai>=0.27.0

